+++ Granada.h
@@ -12,7 +12,7 @@
 
 class Granada {
 public:
-    Granada(sf::Vector2<int> s, sf::Vector2<int> m, int d, sf::Vector2<int> pos);
+    Granada(sf::Vector2<float> s, sf::Vector2<float> m, int d, sf::Vector2<int> pos);
     Granada(const Granada& orig);
     virtual ~Granada();
     
@@ -23,8 +23,8 @@ public:
     int getVelocidad();
     int getRango();
     int getDanyo();
-    sf::Vector2<int> getPosAnterior();
-    sf::Vector2<int> getPosActual();
+    sf::Vector2<float> getPosAnterior();
+    sf::Vector2<float> getPosActual();
     int getEstado();
             
     void setContador(int c);
@@ -35,8 +35,8 @@ public:
     void setVelocidad(int vel);
     void setRango(int r);
     void setDanyo(int g);
-    void setPosAnterior(sf::Vector2<int> v);
-    void setPosActual(sf::Vector2<int> v);
+    void setPosAnterior(sf::Vector2<float> v);
+    void setPosActual(sf::Vector2<float> v);
     
     int updateGranada();
     void pintarGranada(sf::RenderWindow &window);
@@ -51,8 +51,8 @@ protected:
     sf::Sprite* sprite;
     sf::Vector2<float> mov;
     
-    sf::Vector2<int> posAnterior;
-    sf::Vector2<int> posActual;
+    sf::Vector2<float> posAnterior;
+    sf::Vector2<float> posActual;
     sf::Vector2<int> posicionFinal;
     
 private:
